pyproject.toml
[project]
name = "p1reader"
description = "Application to read values on Raspberry Pi Zero W from"
version = "2024.0.1"
requires-python = ">= 3.9"
dependencies = [
    "gunicorn",
    "asyncio",
    "aiohttp",
    "aiopg",
    "psycopg2",
    "pyserial"
]
authors = [
  {name = "Michèl Borgman", email = "mpborgman@hotmail.com"}
]
maintainers = [
      {name = "Michèl Borgman", email = "mpborgman@hotmail.com"}
]
readme = "README.md"
keywords = ["p1", "smart", "meter", "raspberry", "pi", "zero"]
classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 4 - Beta",

  # Indicate who your project is intended for
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",

  # Pick your license as you wish (see also "license" above)
  "License :: OSI Approved :: MIT License",

  # Specify the Python versions you support here.
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9"
]